% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_arrays.R
\name{group_arrays}
\alias{group_arrays}
\title{group CRISPR arrays from a \code{\link{merge_minced}} object by repeat similarity}
\usage{
group_arrays(data)
}
\arguments{
\item{data}{A merged tibble dataset created with \code{\link{merge_minced}}.}
}
\value{
A tibble of CRISPR arrays grouped by repeat consensus sequence.
}
\description{
Given a dataset created with \code{\link{merge_minced}}, this function attempts to
identify CRISPR arrays across individual profiles / samples that have identical
repeat consensus sequences. As the assembly orientation of a CRISPR array
may be different across samples, this function compares reverse complements, too.
}
