% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary.R
\name{group_summary}
\alias{group_summary}
\title{summarize the dataset generated by group_arrays()}
\usage{
group_summary(data)
}
\arguments{
\item{data}{An object created with group_arrays().}
}
\value{
A tibble of CRISPR consensus repeats and the number of samples
in which a repeat is identified.
}
\description{
Given a dataset created with group_arrays(), this function lists
the samples represented by each consensus repeat, sorted by the
number of constituent samples (descending).
}
