% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_minced.R
\name{merge_minced}
\alias{merge_minced}
\title{merge read_minced() objects into a single tibble}
\usage{
merge_minced(..., names = NULL)
}
\arguments{
\item{...}{One or more objects created with read_minced().}

\item{names}{A vector of unique names  to label merged objects. If no names are given, the object names are used.}
}
\value{
A tibble containing the merged data.
}
\description{
Given multiple objects from read_minced(), this function merges
them into a single tibble and creates an ID column from 'names'.
If a vector of names is not supplied, the object names are used as labels.
}
