% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arrays.R
\name{plot_arrays}
\alias{plot_arrays}
\title{Given a dataset created with group_arrays(), plot CRISPR arrays}
\usage{
plot_arrays(
  group,
  cdist = 0.1,
  palette =
    "a6cee3,1f78b4,b2df8a,33a02c,fb9a99,e31a1c,fdbf6f,ff7f00,cab2d6,6a3d9a,ffff99,b15928",
  plot_logo = TRUE
)
}
\arguments{
\item{group}{A subset of the object created by group_arrays() containing only a single consensus repeat.}

\item{cdist}{A numeric scalar in [0,1] representing a proportion of the maximum Levenshtein distance
for a set of spacers. Default = 0.1.}

\item{palette}{A string of hexadecimal color codes used distinguish unique spacers.}

\item{plot_logo}{If FALSE, do not plot the repeat sequence logo. Default = TRUE.}
}
\value{
A plot comparing CRISPR arrays with a shared consensus repeat.
}
\description{
plot_arrays() generates a graphical
representation of array sets associated by shared repeat consensus sequences
(the output of group_arrays())
}
