% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_summary.R
\name{crispr_summary}
\alias{crispr_summary}
\title{summarize CRISPR arrays}
\usage{
crispr_summary(data)
}
\arguments{
\item{data}{An object created with \code{\link{read_minced}}.}
}
\value{
A summary tibble.
}
\description{
Given an object created with \code{\link{read_minced}}, this function
gives the length and number of spacers in each CRISPR array.
}
