% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arrays.R
\name{plot_arrays}
\alias{plot_arrays}
\title{Given a dataset created with group_arrays(), plot CRISPR arrays}
\usage{
plot_arrays(
  group,
  cdist = 0.1,
  palette = default_palette,
  plot_logo = TRUE,
  number_spacers = FALSE
)
}
\arguments{
\item{group}{A subset of the object created by group_arrays() containing only a single consensus repeat.}

\item{cdist}{A numeric scalar in [0,1] representing a proportion of the maximum Levenshtein distance
for a set of spacers. Default = 0.1.}

\item{palette}{A vector of hexadecimal color codes used to distinguish unique spacers.
Default = default_palette, which is equivalent to grDevices::palette.colors(palette = "Polychrome 36")[-c(1:2)]}

\item{plot_logo}{If FALSE, do not plot the repeat sequence logo. Default = TRUE.}

\item{number_spacers}{If TRUE, include cluster number label in each spacer. Default = FALSE.}
}
\value{
A plot comparing CRISPR arrays with a shared consensus repeat.
}
\description{
plot_arrays() generates a graphical
representation of array sets associated by shared repeat consensus sequences
(the output of group_arrays())
}
