% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_minced.R
\name{read_minced}
\alias{read_minced}
\title{Read and transform MinCED data}
\usage{
read_minced(txt, gff)
}
\arguments{
\item{txt}{Path to the minced .txt output}

\item{gff}{Path to the minced .gff output}
}
\value{
A tibble containing the minced data
}
\description{
MinCED is a program used to identify CRISPRs in genomes
or contigs. By default, MinCED outputs both a human-readable
text file and a gff-formatted annotation file. This function
parses both the .txt and .gff outputs of MinCED to create a
single tibble.
}
